Move `CorrelationSurface` data structure back into `computation`
